+++ src/skima.rs
@@ -6,8 +6,8 @@ pub const BUFFER_SIZE: usize = 16 * 1024;
 pub struct SharedData {
     pub flag_server: AtomicU32, // 서버 상태 - 0: 대기, 1: 데이터 전송, 2: 데이터 수신 완료, 3: 종료
     pub flag_client: AtomicU32, // 클라이언트 상태 - 0: 대기, 1: 데이터 전송, 2: 데이터 수신 완료, 3: 종료
-    pub data_len_server_to_client: usize, // 서버->클라이언트 데이터 길이
-    pub data_len_client_to_server: usize, // 클라이언트->서버 데이터 길이
+    pub data_len_server_to_client: u32, // 서버->클라이언트 데이터 길이
+    pub data_len_client_to_server: u32, // 클라이언트->서버 데이터 길이
     pub data_server_to_client: [u8; BUFFER_SIZE], // 서버->클라이언트 데이터
     pub data_client_to_server: [u8; BUFFER_SIZE], // 클라이언트->서버 데이터
 }
